# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server.",
        "title": "My API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/info/client": {
            "get": {
                "description": "Get the client IP address and user agent",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Info"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.InfoClientPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/info/database": {
            "get": {
                "description": "Get the database version",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Info"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.InfoDatabasePayload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/info/version": {
            "get": {
                "description": "Get the golang version",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Info"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.InfoGoVersionPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.InfoClientPayload": {
            "type": "object",
            "properties": {
                "IP": {
                    "type": "string"
                },
                "User-Agent": {
                    "type": "string"
                }
            }
        },
        "models.InfoDatabasePayload": {
            "type": "object",
            "properties": {
                "version": {
                    "type": "string"
                }
            }
        },
        "models.InfoGoVersionPayload": {
            "type": "object",
            "properties": {
                "goversion": {
                    "type": "string"
                }
            }
        },
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Внутренний код ошибки",
                    "type": "integer"
                },
                "data": {
                    "description": "Основные данные для ответа (может быть nil при ошибке)"
                },
                "message": {
                    "description": "Описание ошибки или успеха",
                    "type": "string"
                }
            }
        }
    }
}